# public/swagger/data.yaml
openapi: 3.0.0
info:
  title: 'Galaxy Project OpenApi Documents'
  version: 1.0.0
servers:
  -
    url: '{schema}://{url}:{port}'
    description: 'OpenApi Location'
    variables:
      schema:
        enum:
          - https
          - http
        default: http
      url:
        enum:
          - 127.0.0.1
        default: 127.0.0.1
      port:
        enum:
          - 8888
          - 8000
        default: '8888'
paths:
  /api/v1/auth/token:
    post:
      tags:
        - Auth
        - Token
      summary: 'add new token'
      operationId: auth-token-add
      requestBody:
        description: 后台账号登陆
        content:
          application/json:
            schema:
              required:
                - usr
                - pwd
                - otp_code
              properties:
                usr:
                  type: string
                  example: root
                pwd:
                  description: 客户端需要发送md5
                  type: string
                  example: 200820E3227815ED1756A6B531E7E0D2
                otp_code:
                  type: string
                  example: '348591'
              type: object
      responses:
        '200':
          description: successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/200'
        '429':
          description: 'Rate Limit'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/429'
        '500':
          description: 'Server Internal Error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/500'
      security:
        -
          bearerAuth: []
components:
  schemas:
    '200':
      required:
        - code
        - msg
        - data
      properties:
        code:
          description: 正常返回0，当返回4位数字代表具体业务错误码定义
          type: integer
          default: 0
        msg:
          type: string
          default: successfully
        data:
          type: object
          default: 'null'
      type: object
    '400':
      required:
        - code
        - msg
        - data
      properties:
        code:
          description: 未知请求错误
          type: integer
          default: 1000
        msg:
          type: string
          default: 'something problems'
        data:
          type: object
          default: 'null'
      type: object
    '429':
      required:
        - code
        - msg
        - data
      properties:
        code:
          type: integer
          default: 429
        msg:
          type: string
          default: 'Request Limited'
        data:
          type: object
          default: 'null'
      type: object
    '500':
      required:
        - code
        - msg
        - data
      properties:
        code:
          type: integer
          default: 500
        msg:
          type: string
          default: 'Internal Server Error'
        data:
          type: object
          default: 'null'
      type: object
    '503':
      required:
        - code
        - msg
        - data
      properties:
        code:
          type: integer
          default: 503
        msg:
          type: string
          default: 'Server Busy'
        data:
          type: object
          default: 'null'
      type: object
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  -
    bearerAuth: []